requirements.txt
selenium
webdriver-manager
pytest
python-dotenv